3 projectors & threshold testing
Set up 3 different projectors/cameras in Unity, achieved accurate collision detection between L-Bracket and targets in all three views. Began testing with velocity thresholds; still figuring that part out. We successfully wrote code that detects the velocity in the Z (length of the lab, Y) direction, but we need to figure out what the best threshold for detecting a foot step will be. That, on Thursday!